requirements.txt
astral==3.2
flask==3.1.2
numpy<2
onnxruntime==1.16.0
onvif-zeep==0.2.12
opencv-python-headless==4.10.0.84
piexif==1.1.3
pillow==12.0.0
psutil==7.1.3
python-dotenv==1.2.1
pytz==2025.2
pyyaml==6.0.3
requests==2.32.5
waitress==3.0.2
WSDiscovery==2.1.2
